airtable: Improve URL parsing (#38)
Improve URL parsing in a few ways:
- Add support for parsing record IDs. This makes it possible/easy to
write a workflow like: `Update the "Name" field of an Airtable record
(given by the URL of the record) to a customizable name.`
- Provide a format for the expected input. This seems to help avoid
workflow argument examples that are API URLs like
`https://api.airtable.com`.
- Now that it can optionally parse out IDs, return optional IDs. The
docstring explicitly mentions that the optional values be checked for
non-`None` before use. This helps avoid a first (failing) pass that
fails to type-check because the returned IDs are used without
validating.
- Use `urllib.parse` to parse URLs because... they are URLs. Unquote the
path to handle any encoding shenanigans.
- Rename to `airtable_parse_ids_from_url`. This seems to help code
generation use it correctly, e.g. to either use it for all needed all
IDs or not, as opposed to asking for a URL for the base and separately
asking for a table ID -- I'm not sure why this would be, but that's what
it seemed like to me.

Unrelatedly:
- Format the code.

Note that because of non-determinism, it's hard to have 100% confidence
that this is a strict improvement, but it qualitatively seems a lot
better to me.