airtable: Allow specification of what fields to include (#42)
Allow specification of what fields to include when listing records. This
is a crude way for generated code to assert that the schema of a table
is as it expects.

Airtable tables have field names. When fetching data from Airtable, the
generated code (possibly on its own or at your accidental direction) may
try to access fields with names that do not exist, e.g.:

```python
    records: list[AirtableRecord] = airtable_record_list(base_id, table_id)
    person_values: List[str] = []
    for record in records:
        person_values.append(record.fields["First Name"])
```

where the records do not have a `"First Name"` field and instead have a
`"Name"` field.

This is problematic because the code doesn't encounter an error until it
tries to access the `"First Name"` field. At that moment, it's hard to
provide a useful error message, since it's a plain dict field access
problem, and that (naively) raises a generic key error. It's
subsequently hard for users to work out that it's a schema problem,
especially if it is a small typo, e.g. capitalization error.

With this change, we encourage generated code to specify which fields to
include in the invocation of `airtable_record_list`. In this, it would
hopefully generate:

```python
    records: list[AirtableRecord] = airtable_record_list(base_id, table_id, include_fields={"First Name"})
    person_values: List[str] = []
    for record in records:
        person_values.append(record.fields["First Name"])
```

This code will raise an error in the call to `airtable_record_list`, and
that error will include the schema of the table, so it's both easier for
users to understand what has happened and for the workflow to
(automatically) fix itself.

Closes [#2639](https://github.com/d8e-ai/alda/issues/2639).