Add github issues and comments fetching actions, and a test. (#46)
It's a bit tricky to test these actions because they require an
authenticated wrapper. However, the test mocks / injects a fake for the
symbols / implementation to get things working, and pointing the APIs at
public resources still works for the purposes of a test. Though this
hits a real API so is subject to rate limits from GitHub. But it's okay
for one-off testing (not CI, yet).

Let me know what you think or what alternatives might be worth pursuing!