fix hubspot boolean handling and missing email case (#78)
"distutils" does not exist in Python 3.12, so I switched it to use
Pydantic boolean parsing instead.

I also discovered while testing this that `properties.get("email", "")`
returns `None` for contacts with no email, which causes a pydantic
ValidationError while contstructing the HubSpotContact, so I fixed that
too.

How I tested:
* I added a custom boolean property `test_boolean_property` to contacts
in our HubSpot playground.
* `_coerce_properties_to_lutra`: I made a workflow that exports all the
contacts to Google Sheets with a column for `test_boolean_property`
property, and checked that it exported the values correctly.
* `_coerce_properties_to_hubspot`: I made a workflow that updates
`test_boolean_property` and checked that it worked. (Note: Actually this
doesn't work because the plugin doesn't know that the custom property is
boolean. So to test this, I temporarily changed the check `name in
boolean_property_names` to `name in boolean_property_names or name ==
"test_boolean_property"` so that the coercion logic would get triggered
for that property. Actually fixing this is out of scope of what I was
trying to do, which was just to get the HubSpot plugin working at all
after the 3.12 upgrade.)